feat: Move coffeescript to peer dependencies
Allow to install coffeescript of the particular version used in the project.

BREAKING CHANGE: The package coffeescript was moved to peer dependencies. If you use a package manager with the feature parity of NPM 8, you won't notice a difference, because it will install the peer depednencies automatically. Otherwise you will need to install coffeescript as a dev dependency, like you did karma earlier.